#1 добавлена проверка что соеденяеться одинаковые объекты и перемешёт димострационый проект
